requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --no-hashes
affine==2.4.0
altair==5.5.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6 ; platform_system == 'Windows'
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.0
gitdb==4.0.12
gitpython==3.1.45
idna==3.11
image-similarity-measures==0.3.5
imageio==2.37.2
jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy-loader==0.4
markupsafe==3.0.3
matplotlib==3.10.7
narwhals==2.12.0
networkx==3.6
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
phasepack==1.5
pillow==12.0.0
plotly==6.5.0
protobuf==6.33.1
pyarrow==21.0.0
pydeck==0.9.1
pyfftw==0.15.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
rasterio==1.4.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-image==0.25.2
scipy==1.16.3
six==1.17.0
smmap==5.0.2
streamlit==1.51.0
streamlit-image-comparison==0.0.4
streamlit-image-coordinates==0.4.0
tenacity==9.1.2
tifffile==2025.10.16
toml==0.10.2
tornado==6.5.2
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0 ; platform_system != 'Darwin'
